+++ beast-of-argh/CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add directory for promtail positions
 * Adjust syslog-ng to save remote log files with group read permissions
 * Remove invalid loki and promtail sysrc entries not present in pkg install grafana-loki
+* Change syslog-ng file save perms under options not destination
 
 ---