+++ beast-of-argh/CHANGELOG.md
@@ -7,6 +7,7 @@
 * Remove invalid loki and promtail sysrc entries not present in pkg install grafana-loki
 * Change syslog-ng file save perms under options not destination
 * Correctly set syslog-ng remote logs permissions instead of global option
+* Even more correctly set permissions 0644 on remote logs files
 
 ---